draw user away

+++ app/src/main/java/com/example/vlad/navigation/utils/Vector.java
@@ -11,11 +11,11 @@ import java.util.IllegalFormatException;
 public class Vector {
     private String title;
 
-    private double x;
-    private double y;
-    private double z;
+    private float x;
+    private float y;
+    private float z;
 
-    public Vector(String title, double[] arr) throws IllegalFormatException {
+    public Vector(String title, float[] arr) throws IllegalFormatException {
         if(arr.length < 3) {
             throw new ArraySizeException();
         }
@@ -25,27 +25,27 @@ public class Vector {
         this.title = title;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public double getZ() {
+    public float getZ() {
         return z;
     }
 
-    public void setZ(double z) {
+    public void setZ(float z) {
         this.z = z;
     }
